Turn Mordred missing/error object to NA.

+++ inst/python/Mordred_MD.py
@@ -6,4 +6,4 @@ def getMordredMD_py(smis, ignore_3D = True, thread = 1):
   mols = [Chem.MolFromSmiles(smi) for smi in smis]
   calc = Calculator(descriptors, ignore_3D = ignore_3D)
   res_pd = calc.pandas(mols, nproc = thread)
-  return(res_pd)
+  return(res_pd.astype(float))
